fix(entity-tabs): moving mergeTabs to new util file
